OSN-163. Refactor InitializeContainer with step-based architecture
- Restructured InitializeContainer to use a step-based flow for organization initialization
- Introduced separate components for managing invitations, organization setup, and application redirection
- Extracted common UI elements like Loading, Error, and Title into reusable components
- Simplified the initialization process with a more modular approach
- Maintained existing functionality for handling invitations and organization creation
