+++ ngui/ui/src/components/RecommendationListItemFlavorLabel/RecommendationListItemFlavorLabel.tsx
@@ -1,10 +1,8 @@
 import CloudTypeIcon from "components/CloudTypeIcon";
 import IconLabel from "components/IconLabel";
 
-const RecommendationListItemFlavorLabel = ({ item }) => {
-  return (
-    <IconLabel icon={<CloudTypeIcon fontSize="small" type={item.cloud_type} />} label={item.flavor} alignItems="flex-start" />
-  );
-};
+const RecommendationListItemFlavorLabel = ({ item }) => (
+  <IconLabel icon={<CloudTypeIcon fontSize="small" type={item.cloud_type} />} label={item.flavor} alignItems="flex-start" />
+);
 
 export default RecommendationListItemFlavorLabel;
