+++ ngui/ui/src/containers/InitializeContainer/common/index.tsx
@@ -1,0 +1,5 @@
+import Error from "./Error";
+import Loading from "./Loading";
+import { Title } from "./Title";
+
+export { Title, Loading, Error };
