+++ ngui/ui/src/containers/InitializeContainer/steps/ManageInvitations/index.tsx
@@ -1,0 +1,3 @@
+import StepContainer from "./StepContainer";
+
+export default StepContainer;
