+++ ngui/ui/src/containers/InitializeContainer/steps/ProceedToApplication/index.tsx
@@ -1,0 +1,3 @@
+import ProceedToApplication from "./ProceedToApplication";
+
+export default ProceedToApplication;
